// backend/src/test/recipe/entities/unit.enum.ts
export enum Unit {
  GRAM = 'g',
  KILOGRAM = 'kg',
  MILLILITER = 'ml',
  LITER = 'l',
  TEASPOON = 'tsp',
  TABLESPOON = 'tbsp',
  PIECE = 'pcs',
  CUP = 'cup',
  BALL = 'ball',
  LEAVES = 'leaves',
  MEDIUM = 'medium',
  LARGE = 'large',
  CLOVES = 'cloves',
  LOAF = 'loaf',
  SLICES = 'slices',
  LEMON = 'lemon',
  TABLESPOON_SHORT = 'tbsp',
  TEASPOON_SHORT = 'tsp',
  HEAD = 'head',
  CAN = 'can',
}